refactor(#2605): split filter.rs inline tests into filter_tests.rs (campsite rule)
filter.rs was already 1033 lines — over the 800-line src target — so the 7-line
doc comment justifying `lower_predicate_expr`'s `pub(crate)` visibility tripped
the gate's file-size ratchet. Split by responsibility per epic #1116/#1135,
following this crate's existing `<module>_tests.rs` + `#[path]` convention
(bypass_tests.rs, statics_tests.rs, ...): filter.rs 1033 -> 543, new
filter_tests.rs 505. All 29 filter tests still run, unchanged.

Refs #2605
